+++ source/js/modules/3d/svg/svg-objects-loader.js
@@ -27,5 +27,3 @@ export class SvgObjectsLoader {
     });
   }
 }
-
-
